dedicate http client to separate folder
