+++ src/http/interceptors/index.js
@@ -1,0 +1,7 @@
+import { addRequestInterceptor, addResponseInterceptor } from '../client'
+import { requestWithAuth, handleAuthError } from './auth'
+
+export default function initializeHttpInterceptors() {
+  addRequestInterceptor({ request: requestWithAuth })
+  addResponseInterceptor({ error: handleAuthError })
+}
